requirements.txt
streamlit>=1.28.0
nltk>=3.8.0
google-generativeai>=0.3.0
pandas>=2.0.0
altair>=5.0.0
matplotlib>=3.7.0
numpy>=1.24.0